OnjaroWeb/requirements.txt
flask>=3.0
flask-socketio>=5.3
simple-websocket>=1.0
apscheduler>=3.10
pydantic>=2.0
pyyaml>=6.0
supabase>=2.0
openai>=1.0
httpx>=0.25
beautifulsoup4>=4.12
